% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{check_installed}
\alias{check_installed}
\title{check if package is installed}
\usage{
check_installed(package)
}
\arguments{
\item{package}{}
}
\value{
invisible
}
\description{
check if package is installed
}
\examples{
check_installed("foo")
check_installed("base")
}
