% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_logger}
\alias{test_logger}
\title{Test logger (test utility)}
\usage{
test_logger(start = NULL, end = NULL, msg)
}
\arguments{
\item{start}{test start message}

\item{end}{test end message}

\item{msg}{test message}
}
\value{
message to test output
}
\description{
This is an example test helper function that's also part of the package
namespace. Functions in the \verb{tests/testthat/} folder with a \verb{<helper>.R}
prefix will be automatically loaded.
}
